add password caching support

+++ src/use-cases/remove-password.js
@@ -1,9 +1,13 @@
-export default function makeRemovePassword ({ currentDb }) {
+export default function makeRemovePassword ({ currentDb, currentCache }) {
   return async function removePassword (info) {
-    const result = await currentDb.removePassword(info)
-    if (result == null) {
+    const cacheResult = await currentCache.removePassword(info)
+    if (cacheResult == null) {
+      throw new Error('No cache result')
+    }
+    const dbResult = await currentDb.removePassword(info)
+    if (dbResult == null) {
       throw new Error('No db result')
     }
-    return result
+    return dbResult
   }
 }
